Add item removal button
Created a button for the mountain item to be removed for a particular
user. Upon clicking, the logic looks for authentication in order to
complete the request. Basic structure for a handlebars if statement has
written to display update and remove buttons on the mountain items that
only the logged in user is allowed to update or delete.