Modify handlebars, add style
1. Updated handlebars with #if statements to only show information
about the entry if there is any.
2. Changed around the how the values were displayed, as the item values
didn't flow well. Values that are displayed are now titled with an h5,
and have a <p> section for the information held.
3. Added some structure to the Update button from made in handlebars. It
will show a modal on a click, styled the same as the create entry modal.
4. Some basic style was added to the main it's child 'section', and it's
child class content. Values are to be modified but give more readability
for easing further development.