Fix update functionality
Used 'store' to store the id being passed into the update for from
handlebars. The target id being sent in the update api was set to
`store.mountain.id`. This value was re-defined, and the `.id` was
removed. Fixed the wrong modal from being targeted on an successful
update in order to toggle the form.